Refactor: Standardize API file naming conventions for consistency
