+++ api/src/routes/employees/delete_employee.ts
